Se movió el decorador rol_requerido a un módulo dedicado de autorización. Se hace generica el decorador de autorizacion rol_requerido.
